minor refactoring add proper entry point
