+++ arthas/config.py
@@ -3,7 +3,7 @@ google_api_key = '*'
 telegram_chat_channel = "*"
 telegram_token = "*"
 
-# youtube_channel_id = 'UCbt5BCs0aUDgNwRaUnKtXwA'
-youtube_channel_id = 'UCUyodDg_PnLj885rCgbVN0A'
+youtube_channel_id = 'UCbt5BCs0aUDgNwRaUnKtXwA'    # main
+# youtube_channel_id = 'UCUyodDg_PnLj885rCgbVN0A'  # casino
 
 logger_format = "%(asctime)-15s [%(levelname)5s]: %(message)s"
