+++ pyproject.toml
@@ -17,6 +17,7 @@ opencv-python = "^4.5.5"
 google-api-python-client = "^2.46.0"
 google-auth-oauthlib = "^0.5.1"
 methodtools = "^0.4.5"
+click = "^8.1.3"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
