--- run.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-. /root/venv/bin/activate
-(cd /root/running/; python ../twitchbot/src/main.py)
